new(entity): add main service entities

+++ src/main/java/com/gawlas/CarServiceServer/entities/Client.java
@@ -10,36 +10,36 @@ public class Client extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "pers_id")
-    private Integer persId;
+    private Long persId;
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     public Client() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getPersId() {
+    public Long getPersId() {
         return persId;
     }
 
-    public void setPersId(Integer persId) {
+    public void setPersId(Long persId) {
         this.persId = persId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
